+++ lib/features/settings/ui/pages/flags_page.dart
@@ -28,7 +28,6 @@ class _FlagsPageState extends ConsumerState<FlagsPage> {
     enableMatrixFlag,
     resendAttachments,
     useCompressedJsonAttachmentsFlag,
-    useInboundEventQueueFlag,
     enableHabitsPageFlag,
     enableDashboardsPageFlag,
     enableDailyOsPageFlag,
@@ -61,8 +60,6 @@ class _FlagsPageState extends ConsumerState<FlagsPage> {
         return Icons.refresh_rounded;
       case useCompressedJsonAttachmentsFlag:
         return Icons.compress_rounded;
-      case useInboundEventQueueFlag:
-        return Icons.inbox_rounded;
       case enableHabitsPageFlag:
         return Icons.repeat_rounded;
       case enableDashboardsPageFlag:
@@ -109,8 +106,6 @@ class _FlagsPageState extends ConsumerState<FlagsPage> {
         return context
             .messages
             .configFlagUseCompressedJsonAttachmentsDescription;
-      case useInboundEventQueueFlag:
-        return context.messages.configFlagUseInboundEventQueueDescription;
       case enableHabitsPageFlag:
         return context.messages.configFlagEnableHabitsPageDescription;
       case enableDashboardsPageFlag:
@@ -155,8 +150,6 @@ class _FlagsPageState extends ConsumerState<FlagsPage> {
         return context.messages.configFlagResendAttachments;
       case useCompressedJsonAttachmentsFlag:
         return context.messages.configFlagUseCompressedJsonAttachments;
-      case useInboundEventQueueFlag:
-        return context.messages.configFlagUseInboundEventQueue;
       case enableHabitsPageFlag:
         return context.messages.configFlagEnableHabitsPage;
       case enableDashboardsPageFlag: